+++ src/busses/loads.jl
@@ -168,23 +168,23 @@ end
 return the real and reactive power injections as vectors with 3 phase indices and net.Ntimesteps time
 indices like:
 ```julia
-Pj, Qj = sj_per_unit(my_bus, net)
+sj = sj_per_unit(my_bus, net)
 ...
-Pj[phase][time_step]
+sj[phase][time_step]
 ```
 """
-function sj_per_unit(j::String, net::Network{MultiPhase})::Tuple{Vector{Vector{<:Real}}, Vector{Vector{<:Real}}}
-    Pj = [zeros(net.Ntimesteps) for _ in 1:3] # first dim is phase, like Pj[phs][t]
-    Qj = [zeros(net.Ntimesteps) for _ in 1:3]
+function sj_per_unit(j::String, net::Network{MultiPhase})::Vector{Vector{ComplexF64}}
+    pj = [zeros(net.Ntimesteps) for _ in 1:3] # first dim is phase, like pj[phs][t]
+    qj = [zeros(net.Ntimesteps) for _ in 1:3]
     if j in real_load_busses(net)
         for phs in 1:3
-            Pj[phs] = -net[j, :kws, phs] * 1e3 / net.Sbase
+            pj[phs] = -net[j, :kws, phs] * 1e3 / net.Sbase
         end
     end
     if j in reactive_load_busses(net)
         for phs in 1:3 
-            Qj[phs] = -net[j, :kvars, phs] * 1e3 / net.Sbase
+            qj[phs] = -net[j, :kvars, phs] * 1e3 / net.Sbase
         end
     end
-    return Pj, Qj
+    return pj + im * qj
 end