+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## dev
 - WIP on bus admittance matrix builder from `Network` (called `Ysparse` for now)
+- stop exporting `sj` method
+- change `sj_per_unit(j::String, net::Network{MultiPhase})` to return complex values
 
 ## v0.4.7
 Changes to support documentation in BranchFlowModel.jl (eventually BusInjectionModel.jl too)